[eclipse/xtext#1569] reduced usage of xtend in purexbase
