+++ org.eclipse.xtext.purexbase.ui/build.properties
@@ -1,6 +1,5 @@
 source.. = src/,\
-           src-gen/,\
-           xtend-gen/
+           src-gen/
 bin.includes = META-INF/,\
                .,\
                plugin.xml,\
