--- org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/PureXbaseUiModule.xtend
@@ -1,35 +1,0 @@
-/*
- * generated by Xtext
- */
-package org.eclipse.xtext.purexbase.ui
-
-import org.eclipse.xtend.lib.annotations.FinalFieldsConstructor
-import org.eclipse.xtext.ui.editor.folding.FoldedPosition
-import org.eclipse.xtext.purexbase.ui.autoedit.AutoEditStrategyProvider
-import org.eclipse.xtext.purexbase.ui.PureXbaseUiModule.NoFolding
-import java.util.Collection
-import org.eclipse.xtext.ui.editor.folding.IFoldingRegionProvider
-import org.eclipse.xtext.ui.editor.autoedit.AbstractEditStrategyProvider
-import org.eclipse.xtext.ui.editor.model.IXtextDocument
-import java.util.Collections
-
-/**
- * Use this class to register components to be used within the Eclipse IDE.
- */
-@FinalFieldsConstructor
-class PureXbaseUiModule extends AbstractPureXbaseUiModule {
-	
-	override Class<? extends AbstractEditStrategyProvider> bindAbstractEditStrategyProvider() {
-		return AutoEditStrategyProvider;
-	}
-	
-	def Class<? extends IFoldingRegionProvider> bindFoldingRegionProvider() {
-		return NoFolding;
-	}
-	static class NoFolding implements IFoldingRegionProvider {
-		override Collection<FoldedPosition> getFoldingRegions(
-				IXtextDocument xtextDocument) {
-			return Collections.emptySet();
-		}
-	}
-}
