+++ org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/contentassist/PureXbaseProposalProvider.java
@@ -1,14 +1,12 @@
-/**
+/*
  * generated by Xtext
  */
 package org.eclipse.xtext.purexbase.ui.contentassist;
 
-import org.eclipse.xtext.purexbase.ui.contentassist.AbstractPureXbaseProposalProvider;
 
 /**
  * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#content-assist
  * on how to customize the content assistant.
  */
-@SuppressWarnings("all")
 public class PureXbaseProposalProvider extends AbstractPureXbaseProposalProvider {
 }
