--- org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/contentassist/PureXbaseProposalProvider.xtend
@@ -1,13 +1,0 @@
-/*
- * generated by Xtext
- */
-package org.eclipse.xtext.purexbase.ui.contentassist
-
-import org.eclipse.xtext.purexbase.ui.contentassist.AbstractPureXbaseProposalProvider
-
-/**
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#content-assist
- * on how to customize the content assistant.
- */
-class PureXbaseProposalProvider extends AbstractPureXbaseProposalProvider {
-}
