+++ org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/labeling/PureXbaseDescriptionLabelProvider.java
@@ -1,4 +1,4 @@
-/**
+/*
  * generated by Xtext
  */
 package org.eclipse.xtext.purexbase.ui.labeling;
@@ -8,8 +8,8 @@ import org.eclipse.xtext.xbase.ui.labeling.XbaseDescriptionLabelProvider;
 /**
  * Provides labels for IEObjectDescriptions and IResourceDescriptions.
  * 
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#label-provider
+ * https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#label-provider
  */
-@SuppressWarnings("all")
 public class PureXbaseDescriptionLabelProvider extends XbaseDescriptionLabelProvider {
+
 }
