--- org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/labeling/PureXbaseDescriptionLabelProvider.xtend
@@ -1,25 +1,0 @@
-/*
-* generated by Xtext
-*/
-package org.eclipse.xtext.purexbase.ui.labeling
-
-import org.eclipse.xtext.xbase.ui.labeling.XbaseDescriptionLabelProvider
-//import org.eclipse.xtext.resource.IEObjectDescription
-
-/**
- * Provides labels for IEObjectDescriptions and IResourceDescriptions.
- * 
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#label-provider
- */
-class PureXbaseDescriptionLabelProvider extends XbaseDescriptionLabelProvider {
-
-	// Labels and icons can be computed like this:
-	
-//	override text(IEObjectDescription ele) {
-//		ele.name
-//	}
-//	 
-//	override image(IEObjectDescription ele) {
-//		ele.EClass.name + '.gif'
-//	}	 
-}
