+++ org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/labeling/PureXbaseLabelProvider.java
@@ -1,4 +1,4 @@
-/**
+/*
  * generated by Xtext
  */
 package org.eclipse.xtext.purexbase.ui.labeling;
@@ -12,10 +12,11 @@ import org.eclipse.xtext.xbase.ui.labeling.XbaseLabelProvider;
  * 
  * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#label-provider
  */
-@SuppressWarnings("all")
 public class PureXbaseLabelProvider extends XbaseLabelProvider {
-  @Inject
-  public PureXbaseLabelProvider(final AdapterFactoryLabelProvider delegate) {
-    super(delegate);
-  }
+
+	@Inject
+	public PureXbaseLabelProvider(AdapterFactoryLabelProvider delegate) {
+		super(delegate);
+	}
+
 }
