--- org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/labeling/PureXbaseLabelProvider.xtend
@@ -1,21 +1,0 @@
-/*
-* generated by Xtext
-*/
-package org.eclipse.xtext.purexbase.ui.labeling
-
-import com.google.inject.Inject
-import org.eclipse.emf.edit.ui.provider.AdapterFactoryLabelProvider
-import org.eclipse.xtext.xbase.ui.labeling.XbaseLabelProvider
-
-/**
- * Provides labels for EObjects.
- * 
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#label-provider
- */
-class PureXbaseLabelProvider extends XbaseLabelProvider {
-
-	@Inject
-	new(AdapterFactoryLabelProvider delegate) {
-		super(delegate);
-	}
-}
