+++ org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/outline/PureXbaseOutlineTreeProvider.java
@@ -1,4 +1,4 @@
-/**
+/*
  * generated by Xtext
  */
 package org.eclipse.xtext.purexbase.ui.outline;
@@ -8,12 +8,11 @@ import org.eclipse.xtext.ui.editor.outline.impl.DefaultOutlineTreeProvider;
 
 /**
  * Customization of the default outline structure.
- * 
+ *
  * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#outline
  */
-@SuppressWarnings("all")
 public class PureXbaseOutlineTreeProvider extends DefaultOutlineTreeProvider {
-  public boolean _isLeaf(final Model m) {
-    return true;
-  }
+	public boolean _isLeaf(Model m) {
+		return true;
+	}
 }
