--- org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/outline/PureXbaseOutlineTreeProvider.xtend
@@ -1,19 +1,0 @@
-/*
-* generated by Xtext
-*/
-package org.eclipse.xtext.purexbase.ui.outline
-
-import org.eclipse.xtext.purexbase.pureXbase.Model
-import org.eclipse.xtext.ui.editor.outline.impl.DefaultOutlineTreeProvider
-
-/**
- * Customization of the default outline structure.
- *
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#outline
- */
-class PureXbaseOutlineTreeProvider extends DefaultOutlineTreeProvider {
-
-	def _isLeaf(Model m) {
-		true
-	}	
-}
