+++ org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/quickfix/PureXbaseQuickfixProvider.java
@@ -1,0 +1,26 @@
+/*
+ * generated by Xtext
+ */
+package org.eclipse.xtext.purexbase.ui.quickfix;
+
+import org.eclipse.xtext.xbase.ui.quickfix.XbaseQuickfixProvider;
+
+/**
+ * Custom quickfixes.
+ *
+ * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#quick-fixes
+ */
+public class PureXbaseQuickfixProvider extends XbaseQuickfixProvider {
+
+//	@Fix(PureXbaseValidator.INVALID_NAME)
+//	public void capitalizeName(final Issue issue, IssueResolutionAcceptor acceptor) {
+//		acceptor.accept(issue, "Capitalize name", "Capitalize the name.", "upcase.png", new IModification() {
+//			public void apply(IModificationContext context) throws BadLocationException {
+//				IXtextDocument xtextDocument = context.getXtextDocument();
+//				String firstLetter = xtextDocument.get(issue.getOffset(), 1);
+//				xtextDocument.replace(issue.getOffset(), 1, firstLetter.toUpperCase());
+//			}
+//		});
+//	}
+
+}
