--- org.eclipse.xtext.purexbase.ui/src/org/eclipse/xtext/purexbase/ui/quickfix/PureXbaseQuickfixProvider.xtend
@@ -1,27 +1,0 @@
-/*
-* generated by Xtext
-*/
-package org.eclipse.xtext.purexbase.ui.quickfix
-
-import org.eclipse.xtext.xbase.ui.quickfix.XbaseQuickfixProvider
-//import org.eclipse.xtext.ui.editor.quickfix.Fix
-//import org.eclipse.xtext.ui.editor.quickfix.IssueResolutionAcceptor
-//import org.eclipse.xtext.validation.Issue
-
-/**
- * Custom quickfixes.
- *
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#quick-fixes
- */
-class PureXbaseQuickfixProvider extends XbaseQuickfixProvider {
-
-//	@Fix(MyDslValidator.INVALID_NAME)
-//	def capitalizeName(Issue issue, IssueResolutionAcceptor acceptor) {
-//		acceptor.accept(issue, 'Capitalize name', 'Capitalize the name.', 'upcase.png') [
-//			context |
-//			val xtextDocument = context.xtextDocument
-//			val firstLetter = xtextDocument.get(issue.offset, 1)
-//			xtextDocument.replace(issue.offset, 1, firstLetter.toUpperCase)
-//		]
-//	}
-}
