--- org.eclipse.xtext.purexbase.ui/xtend-gen/org/eclipse/xtext/purexbase/ui/PureXbaseUiModule.java
@@ -1,42 +1,0 @@
-/**
- * generated by Xtext
- */
-package org.eclipse.xtext.purexbase.ui;
-
-import java.util.Collection;
-import java.util.Collections;
-import org.eclipse.ui.plugin.AbstractUIPlugin;
-import org.eclipse.xtend.lib.annotations.FinalFieldsConstructor;
-import org.eclipse.xtext.purexbase.ui.AbstractPureXbaseUiModule;
-import org.eclipse.xtext.purexbase.ui.autoedit.AutoEditStrategyProvider;
-import org.eclipse.xtext.ui.editor.autoedit.AbstractEditStrategyProvider;
-import org.eclipse.xtext.ui.editor.folding.FoldedPosition;
-import org.eclipse.xtext.ui.editor.folding.IFoldingRegionProvider;
-import org.eclipse.xtext.ui.editor.model.IXtextDocument;
-
-/**
- * Use this class to register components to be used within the Eclipse IDE.
- */
-@FinalFieldsConstructor
-@SuppressWarnings("all")
-public class PureXbaseUiModule extends AbstractPureXbaseUiModule {
-  public static class NoFolding implements IFoldingRegionProvider {
-    @Override
-    public Collection<FoldedPosition> getFoldingRegions(final IXtextDocument xtextDocument) {
-      return Collections.<FoldedPosition>emptySet();
-    }
-  }
-  
-  @Override
-  public Class<? extends AbstractEditStrategyProvider> bindAbstractEditStrategyProvider() {
-    return AutoEditStrategyProvider.class;
-  }
-  
-  public Class<? extends IFoldingRegionProvider> bindFoldingRegionProvider() {
-    return PureXbaseUiModule.NoFolding.class;
-  }
-  
-  public PureXbaseUiModule(final AbstractUIPlugin plugin) {
-    super(plugin);
-  }
-}
