--- org.eclipse.xtext.purexbase.ui/xtend-gen/org/eclipse/xtext/purexbase/ui/quickfix/PureXbaseQuickfixProvider.java
@@ -1,15 +1,0 @@
-/**
- * generated by Xtext
- */
-package org.eclipse.xtext.purexbase.ui.quickfix;
-
-import org.eclipse.xtext.xbase.ui.quickfix.XbaseQuickfixProvider;
-
-/**
- * Custom quickfixes.
- * 
- * See https://www.eclipse.org/Xtext/documentation/310_eclipse_support.html#quick-fixes
- */
-@SuppressWarnings("all")
-public class PureXbaseQuickfixProvider extends XbaseQuickfixProvider {
-}
